+++ WarsztatAPI/Controllers/VehicleController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -16,25 +17,43 @@ namespace WarsztatAPI.Controllers
         [HttpGet("GetAll")]
         public ActionResult GetAll()
         {
+
+            try
             {
-                try
-                {
-                    Vehicle[] results = MySqlConnector.ExecuteQueryResult<Vehicle>($"Select vehicles.*,fuel_types.Name from vehicles join fuel_types on vehicles.Fuel_Id=fuel_types.Id group by Vehicles.Id;");
-                    return Ok(results);
-                }
-                catch (Exception e)
-                {
-                    return BadRequest(e.Message);
-                }
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
+                Vehicle[] results = MySqlConnector.ExecuteQueryResult<Vehicle>($"Select vehicles.*,fuel_types.Name from vehicles join fuel_types on vehicles.Fuel_Id=fuel_types.Id group by Vehicles.Id;");
+                return Ok(results);
             }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
+            }
+            catch (Exception e)
+            {
+                return BadRequest(e.Message);
+            }
+
         }
         [HttpGet("Get")]
         public ActionResult Get([FromHeader] uint id)
         {
             try
             {
-                Vehicle[] results = Filter("vehicles.Id",id);
-                return results.Length>0 ? Ok(results[0]) : BadRequest("Brak pojazdu");
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
+                Vehicle[] results = Filter("vehicles.Id", id);
+                return results.Length > 0 ? Ok(results[0]) : BadRequest("Brak pojazdu");
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -47,8 +66,17 @@ namespace WarsztatAPI.Controllers
         {
             try
             {
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
                 return Ok(Filter("Brand", brand));
-                
+
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -60,7 +88,16 @@ namespace WarsztatAPI.Controllers
         {
             try
             {
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
                 return Ok(Filter("Model", model));
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -68,11 +105,20 @@ namespace WarsztatAPI.Controllers
             }
         }
         [HttpGet("Filter/ProductionYear")]
-        public ActionResult FilterProductionYear([FromHeader]uint min, [FromHeader] uint max)
+        public ActionResult FilterProductionYear([FromHeader] uint min, [FromHeader] uint max)
         {
             try
             {
-                return Ok(FilterRange("Production_Year", min,max));
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
+                return Ok(FilterRange("Production_Year", min, max));
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -84,8 +130,17 @@ namespace WarsztatAPI.Controllers
         {
             try
             {
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
                 Vehicle[] results = Filter("Vin", vin);
                 return results.Length > 0 ? Ok(results[0]) : BadRequest("Brak pojazdu");
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -94,11 +149,20 @@ namespace WarsztatAPI.Controllers
         }
 
         [HttpGet("Filter/EnginePower")]
-        public ActionResult FilterEnginePower([FromHeader] uint min,[FromHeader] uint max)
+        public ActionResult FilterEnginePower([FromHeader] uint min, [FromHeader] uint max)
         {
             try
             {
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
                 return Ok(FilterRange("Engine_Power", min, max));
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -110,7 +174,16 @@ namespace WarsztatAPI.Controllers
         {
             try
             {
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
                 return Ok(FilterRange("Engine_Capacity", min, max));
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -122,18 +195,27 @@ namespace WarsztatAPI.Controllers
         {
             try
             {
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
                 return Ok(Filter("Fuel_Id", fuelType));
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
                 return BadRequest(e.Message);
             }
         }
-        private static Vehicle[] Filter(string col,object var)
+        private static Vehicle[] Filter(string col, object var)
         {
-            return MySqlConnector.ExecuteQueryResult<Vehicle>($"Select vehicles.*,fuel_types.Name from vehicles join fuel_types on vehicles.Fuel_Id=fuel_types.Id where {col} = '{var}' group by Vehicles.Id;");          
+            return MySqlConnector.ExecuteQueryResult<Vehicle>($"Select vehicles.*,fuel_types.Name from vehicles join fuel_types on vehicles.Fuel_Id=fuel_types.Id where {col} = '{var}' group by Vehicles.Id;");
         }
-        private static Vehicle[] FilterRange(string col, object min,object max)
+        private static Vehicle[] FilterRange(string col, object min, object max)
         {
             return MySqlConnector.ExecuteQueryResult<Vehicle>($"Select vehicles.*,fuel_types.Name from vehicles join fuel_types on vehicles.Fuel_Id=fuel_types.Id where {col} >= '{min}' and {col}<={max} group by Vehicles.Id;");
         }
@@ -152,11 +234,19 @@ namespace WarsztatAPI.Controllers
             }
         }
         [HttpPut("Update/Brand")]
-        public ActionResult UpdateBrand([FromHeader]uint id,[FromHeader] string brand)
+        public ActionResult UpdateBrand([FromHeader] uint id, [FromHeader] string brand)
         {
             try
             {
-                return Update(id,"Brand", brand);
+                return Update(id, "Brand", brand);
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -170,6 +260,14 @@ namespace WarsztatAPI.Controllers
             {
                 return Update(id, "Model", model);
             }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
+            }
             catch (Exception e)
             {
                 return BadRequest(e.Message);
@@ -181,6 +279,14 @@ namespace WarsztatAPI.Controllers
             try
             {
                 return Update(id, "Production_Year", productionYear);
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -194,6 +300,14 @@ namespace WarsztatAPI.Controllers
             {
                 return Update(id, "Vin", vin);
             }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
+            }
             catch (Exception e)
             {
                 return BadRequest(e.Message);
@@ -205,6 +319,14 @@ namespace WarsztatAPI.Controllers
             try
             {
                 return Update(id, "Registration_Number", registrationNumber);
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -218,6 +340,14 @@ namespace WarsztatAPI.Controllers
             {
                 return Update(id, "Engine_Power", enginePower);
             }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
+            }
             catch (Exception e)
             {
                 return BadRequest(e.Message);
@@ -229,6 +359,14 @@ namespace WarsztatAPI.Controllers
             try
             {
                 return Update(id, "Engine_Capacity", engineCapacity);
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
@@ -242,6 +380,14 @@ namespace WarsztatAPI.Controllers
             {
                 return Update(id, "Fuel_Id", fuelType);
             }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
+            }
             catch (Exception e)
             {
                 return BadRequest(e.Message);
@@ -251,6 +397,8 @@ namespace WarsztatAPI.Controllers
 
         ActionResult Update(uint id, string col, object var)
         {
+            var token = JwtService.Verify(Request.Cookies["jwt"]);
+            if (token.Claims.First(x => x.Type == "IsSuperUser" && x.Value == true.ToString()) is null) return Unauthorized();
             return MySqlConnector.ExecuteNonQueryResult($"update vehicles set {col}='{var}' where Id = {id}")
                 > 0 ? Ok("Pomyślnie edytowano") : BadRequest("Coś poszło nie tak");
         }
@@ -260,8 +408,18 @@ namespace WarsztatAPI.Controllers
         {
             try
             {
+                var token = JwtService.Verify(Request.Cookies["jwt"]);
+                if (token.Claims.First(x => x.Type == "IsSuperUser" && x.Value == true.ToString()) is null) return Unauthorized();
                 return MySqlConnector.ExecuteNonQueryResult($"delete from vehicles where Id = {id}")
                     > 0 ? Ok("Pomyślnie usunięto") : BadRequest("Brak pojazdu");
+            }
+            catch (ArgumentNullException)
+            {
+                return Unauthorized("Nie jesteś zalogowany");
+            }
+            catch (SecurityTokenExpiredException)
+            {
+                return Unauthorized("Przekroczono czas sesji");
             }
             catch (Exception e)
             {
